openembedded-lkft: Skip LTP commands test
[Arnd] Looked into this and we should disable [commands] because it only tests
if user-space is compliant and doesn’t actually do anything to test the kernel.
Failing test cases all have no dependency on the kernel. Concerned about the
overhead of running the commands test as it will take some time. Let’s mark
them as uninteresting at a minimum, but recommends disabling it.

Change-Id: I80bfbb6e4d310b2d46dd09697a63027572313419
Signed-off-by: Dan Rue <dan.rue@linaro.org>