zephyr-upstream: discover test list dynamically
The problem of hardcoded test list is that not all tests are build for
all devices. For example BT tests are not suitable on qemu.

To solve the above issue, this patch generates test list by finding all
the zephyr.bin files in out/${PLATFORM}/tests directory. Some of the
tests don't produce usable output, they are excluded.

Change-Id: Icb1468423009fadcd1e39f54007a9462fa76aaa7
Signed-off-by: Chase Qi <chase.qi@linaro.org>