rpb-openembedded: generate output of repo diffmanifests
repo diffmanifests, when run in a repo workspace, and with an old and new manifest, will
list for each project all commits changes between the old and the new state.

So let's push in the CI artifacts the pinned-manifest.xml file, so that it can
be used in next build to generate the manifest differences. There is a bit of
logic in the scripts to handle situation where the pinned-manifest.xml does not
exist (such as the first time the build will be run again, or in case of any
infrastructure error).

Change-Id: Ia339e9126281910666a9be8b77a22344a0de67f7
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>