openembedded-armv8-ilp32: make kernel verbose, add resize-helper
In order for LAVA to find its place in the boot process, a
non-quiet kernel is required. This is achieved by requesting
the CMDLINE to remove the quiet particle in local.conf.

A few tests in LAVA do require space available for their
scratch, so the resize-helper utility comes in handy.

Change-Id: I35c1c8a3feaa2bf7bc0bb57b339a067ca3e48813
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>