optee-build: juno: generate a recovery image to flash in NOR flash
ftp-upload isn't supported by LAVA. In order to flash our own bl1/fip
binaries, we need to use the recovery image to flash in NOR flash.

Leave a commented section in case we need to adjust to give more space for
a larger kernel or ramdisk. Right now, kernel is 15M and ramdisk is 10.1M.
They fit in the default image configuration.

Change-Id: I574d83f242ee3aefedabca6065edffac3f4c4ad3
Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>