lt-qcom-openembedded-rpb: Add own LAVA jobs for Wifi and BT tests
The LAVA dispatcher support for reboot between a test session is device
specific (boot process related). For example in db410c when reboots it
will enter in fastboot mode causing a lava-test-shell timeout.

The WiFi and BT tests needs to be executed after boot because it looks
at the initial status of the devices and could be contaminated for other
tests. This also makes space for other tests related to WiFi and BT.

A bad thing is now the download/flash image process needs to be executed
tree times.

Change-Id: I4ffa676008d3ec3807765a6917445dcd26777ef6
Signed-off-by: Aníbal Limón <anibal.limon@linaro.org>