zephyr-upstream: build specifically commit that triggered it
A timing issue was made evident with the extensive times it
takes to go from the first to the last job in the matrix
configuration: by the time the last job's turn is up, the Git
repository may have changed already, and thus a different
codebase would be built.

To fix this, anchor against the GIT_REVISION that triggered
the pipeline, by passing that revision as a parameter to the
downstream jobs.

Change-Id: I66f7ebff3e4b12da9280e4c56bde468c0f1e3659
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>