openembedded-lkft: juno: mount /dev/sda1 onto /ltp-tmp and use it as TMPDIR for LTP testing
There is a bunch of LTP test failed or skipped on Juno because of
NFS mounted file system.

This patch mounts /dev/sda1 onto /ltp-tmp which is TMPDIR for LTP testing.
The total number of pass tests increased from 944 to 979 outof 1100.

If mount failes test job would be aborted as incomplete by the help
of LAVA API "lava-test-raise"

Ref:
----
ltp: update test definition to allow for TMPDIR settings
https://projects.linaro.org/browse/QA-1723

Change-Id: I7c11a4174ca3a59e07def13581b8734c4b7c709d
Signed-off-by: Naresh Kamboju <naresh.kamboju@linaro.org>