openembedded-lkft: set kselftests version for mainline and next
This patch changes the version of kselftests executed on mainline and
next kernels to the one that comes from the same source as the kernel
itself. All other kernels remain on the 'current stable' version of
kselftests.

Version of the mainline and next kselftests is reported as 'make
kernelversion'. The URL points to the git repository URL and revision is
set to git commit ID that the kernel was built from.

Change-Id: Iedf0d0f24ec8b0f9987215d4e6a0c551f7886e84
Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>