linaro-hikey-stable-rc-4.4-merge-and-tag: differentiate multiple tags on same day
Some times, the same 4.4 RC branch might be updated on the
very same day before it can be considered final (as final as
an RC can be deemed), and the job would fail because the tag
upon the merge has already been employed.

This change introduces an unique suffix (BUILD_NUMBER)for the tag in
case there is more than one on the same day.

This is also compatible with other jobs that relying on
sorting of tags using `sort -V`.

Change-Id: I15c2166f62ce619a161986af05cf5e7b268262f3
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>