openembedded-lkft: standardize on using PUB_DEST
The publishing destination is now exposed via a job string parameter.
It's the best practice to use the environment variable PUB_DEST to store
its value.
Rename SNAPSHOTS_URL to BASE_URL, which also makes it more consistent with
OE RPB jobs.
Adjust the relevant jobs, scripts and templates to match the new behavior.
Get rid of the switch/case in the common builders script, since it is a
per build configuration and not a shared common code.

Change-Id: If8aa200aab44c71328927f1d131d40089c18197b
Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>