leg-kolla: failure is also success
We are building over 200 images. Some of them are known to fail, other
ones can sometimes fail. It is fine.

So let not break on failure but continue build to the end. And give us
logs so we can check what failed and why so fixes can be prepared and
sent.

Change-Id: Ie1314d30d47ee5cd75b46b0ef1d0537d16a683e4