openembedded-lkft: install python deps in virtualenv
In order to make sure that all dependencies are installed,
submit_for_testing.sh uses virtualenv. The venv is deleted after the
script completes.

This patch has a drawback that virtualenv is created each time
submit_for_testing.sh is executed. This happens for each device defined
in the lava-job-definitions/devices

Change-Id: I0331166828ad05f9d96979488d03ff91aed18042
Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>