rpb-openembedded/builders.sh: use regular image for LAVA
The LAVA console image is too big to fit in X15's. Jobs
currently fail because of that.

Since this image is solely used for boot testing for the
time being, we can resort to the regular console image
while the LAVA image makes progress in its new year's
resolution of being more fit.

Change-Id: I2169bf291526faeb997e7594f8c5ef73ce7061cb
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>