lt-qcom: Add lt-qcom-linux-integration build
This build is aim to generate an Android boot image using a pre-compiled
kernel image, if the kernel modules are specified it will be copied into
the rootfs.

The basic idea is to make the Linux kernel integration testing
using a minimal initramfs image (rootfs), a jenkins trigger will be
implemented to monitor a precompiled kernel images (kernel ci) and then
if detect a new one starts this build with the last initramfs image
available.

The script support ext4/img/cpio rootfs images if the rootfs is cpio it will be
used as a ramdisk in the Android boot image.

The builders-deps.sh script is separated to don't enforce deps installation
when run manually builders.sh (other distros).

Change-Id: Ibb7a292db66f0b1d66eb2e84c12ba30b3f814724
Signed-off-by: Aníbal Limón <anibal.limon@linaro.org>