LKFT: LTP: Use skipfile-lkft.yaml
Pass board (DEVICE_TYPE), branch (KERNEL_VERSION), and environment names
to ltp, and it will generate a skiplist from skipfile-lkft.yaml.

This introduces the concept of 'environment', which is defaulted to
'production' but set to 'staging' in such a case.

Introduce a variable "KERNEL_VERSION_OVERRIDE" for the cases, as in
mainline and next, where KERNEL_VERSION isn't suitable from a skiplist
context.

Change-Id: I573b8f659e936d307423b1d45d6aa744ee2b222d
Signed-off-by: Dan Rue <dan.rue@linaro.org>