LKFT: Staging: Fix KERNEL_VERSION to be consistent
In production, KERNEL_VERSION is a branch name, i.e. '4.14', '4.9', etc.
In the case of next and mainline, it is set to 'git' and
KERNEL_VERSION_OVERRIDE is set to 'next' or 'mainline', respectively.

Thinking that 'explicit is better than implicit', use a simple case
statement rather than trying to be clever with deduction.

The consequence of this bug was that skipfiles are potentially not being
generated correctly in staging, due to the result of 'make
kernelversion' being passed into skipgen as the branch argument.

Untested. Please spot-check for errors.

Change-Id: Iecf74868766be9b279ad48024969819c14449167
Signed-off-by: Dan Rue <dan.rue@linaro.org>