lt-qcom-linux-automerge: Add buildset to do linux integration automerge
Uses automerge ci tool [1], the configuration could be stored inside
a configuration repository/branch, if not a default one is used only
with baseline.

This job support persistent storage for kernel upstream repository to avoid
clone every time [2].

After merge the kernel integration repo will be build for arm and arm64,
if both builds are succeed then push to a repo that Kernel CI looks for
testing.

For now my github repo [3] is used for automerge while merge process is
made on upstream repo [1].

[1] https://git.linaro.org/power/automerge.git/
[2] git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
[3] https://github.com/alimon/automerge.git

Change-Id: I4a70845df858aceeeaaecd00cb6101949fc4e5f5
Signed-off-by: Aníbal Limón <anibal.limon@linaro.org>