LKFT: Allow configs repo override in staging
Modify submit_for_testing.sh to allow the configs repo url to be
overridden using the variable named 'CONFIGS_REPO_URL_OVERRIDE'. If the
variable isn't set, the default behavior remains the same.

Likewise, allow a git revision to be specified using variable
'CONFIGS_REPO_REV_OVERRIDE'. When not specified, the 'master' branch is
used, as is the current default behavior.

These changes are then allowed to be used as parameters in the
lkft-staging-no-skiplist job so that we can test changes to the configs
repo (especially lava template changes).

Change-Id: I219d19f95e583805e072fd56b0dda1e22c2bf7ab
Signed-off-by: Dan Rue <dan.rue@linaro.org>