lkft-hikey-aosp: run syscall test with reboot support
This patch depends on this patch:
https://git.linaro.org/qa/test-definitions.git/commit/automated/android/tradefed?id=806d42e4140b721422d0b0cd01f924042f9fde5e

We are enabling the monitor_fastboot script.
The script basically runs fastboot boot boot.img
which waits for device to be in fastboot mode.

While running the test, the test reboots the device
and this is when the script would boot the device back
to userspace and the test continues to execute.

This is needed to be done only because the device in LAVA
have the configuration where device waits in fastboot mode
in bootloader.

These are proposed solution to avoid the issue but need a lot of work:
1. Relay for hikey to unshort 5 and 6 while running the test which allows the device to get back to userspace at reboot.
2. LAVA provide some mechanism to send messages from test to UART port on the device.

Change-Id: Ib33bb438065db4f69cbe176845de33f01f46fcd0
Signed-off-by: Vishal Bhoj <vishal.bhoj@linaro.org>