LKFT: qemu_arm: Use zImage rather than uImage
qemu_arm jobs are not booting, due to the kernel type (uImage) being
used.

When running qemu_arm jobs, KERNEL_URL which is built from KERNEL_IMG
needs to use the zImage rather than the uImage. In such a case,
explicitly retrieve the zImage name.

It appears that the x15 device itself does not use KERNEL_IMG or
KERNEL_URL (it only uses the rootfs file), so this should not impact
existing x15 jobs.

Change-Id: I37e787905737464488b8e04e8aa054d4bed43547
Signed-off-by: Dan Rue <dan.rue@linaro.org>