zephyr: separate arm/master output directories
If a build is triggered for the arm branch and anotherfor the
master branch simultaneously, AND if the matrix combination
also happens at the very same time, one build might interfere
with the other.

To accomodate both builds, we can simply move the output of
each build to its own directory.

Change-Id: I1c1ae93d78d69730b1932c623a289435d67a6bd4
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>