rpb-openembedded: update LAVA boot template to hi6220-hikey-r2
With the new version of HiKey firmware there is also a new
method for booting up. Instead of the old:
  method: uefi-menu
  commands: fastboot
we now have:
  method: grub
  commands: installed

A very important change is adding this in the context of the
job definition:
  context:
    kernel_start_message: "EFI stub: Using DTB from configuration table"

That will allow LAVA to identify the boot up sequence of the
RPB builds, which currently include "quiet" in the kernel
command line.

The ptable was also changed to match what the health check
uses. Whereas before it was:
  reference-platform/embedded/hikey/16.12/
we will now use:
  reference-platform/components/uefi-staging/49/

Change-Id: I71e9e79a67621a548383adeee4153df29621c345
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>