lt-qcom-bootloader/dragonboard820c: creates gpt_both*.bin files
Fixes dd0c503b979f (lt-qcom-bootloader: dragonboard820c: update GPT and UFS
provisioning files). Since we migrated to our own gpt_main/backup files, we
missed the gpt_both files which are needed to update GPT on the device using
fastboot flash partition command. The gpt_both files are just the concatenation
of the primary and backup GPT header, so it's safe to generate them on the fly.

Change-Id: I414227b914ea192e1c353bde2d1c69418bb61823
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>