openembedded-lkft: make metadata optional for LKFT LAVA templates
In case of generic-build.sh some expected metadata env variables are not
defined. This causes LAVA template generation to fail making the whole
build failed. Since the environment variables are usually LKFT specific
and templates are re-used in other projects, making part of the metadata
section optional. In case variables are not present, value 'unknown'
will be used.

Change-Id: I87c5eb7d232ef6f8eca455ec1c68fb934dcfcc75
Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>