erp-test-automation: add matrix project
The existing erp-test-automation.yaml freestyle project is simple but
has a potential issue. When it fails to provision or test one host, the
related command will return non-zero exit code and ask for retry. As we
have '-e' enabled in build script, non-zero exit will mark the job as
failure and kill the job. This breaks the provision or testing on other
hosts.

It is possible to capture the exit code and decide what to do for the
next step. In this way, we have to amend host group dynamically and pass
it to test status monitor action. It is more complex then using matrix
build. With matrix build, hosts will be separated completely.

Change-Id: Idad8b006e053d5d288c121442a905b669b815268
Signed-off-by: Chase Qi <chase.qi@linaro.org>