rpb-openembedded/lava-job-definitions: fix timeout for Juno
Juno has been failing for quite some time due to a change in
the way LAVA handles timeouts. Whereas previously it was
`u-boot-interrupt' handling that timeout, since an update to
LAVA on February 9 (2018.2) that token changed to
`bootloader-interrupt'.

Setting that timeout to 120 seconds (instead of the default of
30 seconds), as originally intended, gives more time to allow
for a complete flashing of the board, thus letting the builds
boot again.

This was discovered and fixed by Victor Chong:
  https://github.com/96boards/meta-96boards/pull/252#issuecomment-398787962

Change-Id: I5e6a093918b80d6b4d4469cd43fdb13bfd15e3e5
Suggested-by: Victor Chong <victor.chong@linaro.org>
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>