erp-test-automation-matrix: fix race condition for image upload
Image upload is enabled by default in ansible role. It only upload image
when it is not exist yet in mr-provisioner's database. However, matrix
build erp-test-automation-matrix run the same ansible playbook on
multiple platforms in parallel almost in the same time, means image
existence query responds for all of these threads will be 'not exist
yet', which causes that the same image will be uploaded up to the number
of platforms times. To break the race condition, this patch add 10
minutes delay before the execution of ansible playbook for all platforms
except d05. In this way, images will be upload by d05 and other
platforms will know the images already exist and skip upload step.

Change-Id: Icda45d6b1c07bd698e22169255f7b768f488936d
Signed-off-by: Chase Qi <chase.qi@linaro.org>