tcwg-regression-detection: remove heavy triggers
Triggering regression detection on every commit to master in
jenkins-scripts and abe repo is heavy on resource usage. It is also not
necessary since regression is already done when updating tested
branches (via tcwg-abe-tested job) so is really just a convenience. In
all, it is not a worthwhile use of our resources so this patch removes
these trigger.

Change-Id: I713f868d73e6e247088261b628d64edc66571e89