android-lcr hikey builds: support vendor partition for master builds
by utilizing one environment variant ANDROID_VERSION_SUFFIX
to add the value of it to the end of the default template name.
And we could use this ANDROID_VERSION_SUFFIX for other templates as well when necessary.

Also updated related files to support test_submit_for_testing.sh script work

Change-Id: Ie4ab5e9c4b34483897c15df988097a0b83eb33fc
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>