openembedded-lkft: temporary workaround for Glibc and Rocko
A patch is required for Glibc in order to compile eBPF
programs using Clang. This only manifests with Rocko as it's
better at isolating the building environment from the host.

Patch has been sent upstream for discussion [1], but it looks
promising upon initial testing. This temporary workaround
must be removed once there is a conclusion on that patch.

[1] https://patchwork.openembedded.org/series/13090/

Change-Id: I478e70f5c0d3174e4bcda7e342eaa4011afa8206
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>