Fix tcwg-regression-detection failure
Quote removal is not performed in here-document so the setting of
extra_options for dev build testing will lead to that variable having
quotes in it. That later trips the eval in tcwg-dev-build because the
second evalation will execute "abe_branch=foo distro=bar" rather than
abe_branch=foo distro=bar. The former is interpreted as a command to
execute while the latter is 2 variables being set.

This patch removes the quote from the here-document setting
extra_options as it is not necessary.

Change-Id: If0a5cc7ad159a5ebb835a4093d97baf85ccd5052