openembedded-lkft: add TEST_SUITES parameter
The TEST_SUITES parameter controls what test suites to run
through qa-reports. It can generally be one of these:
* kselftests
* libhugetlbfs
* ltp
or, alternatively:
* all
* sanity
* none
or even blank, which will be translated as "all", or just the
name of the specific test suite, like:
* ltp-syscalls

The "sanity" test suite is implied with "all". Specifying just
"sanity" will run those test plans and nothing else.

Change-Id: I4e0a34dce32622fa995ef57d74b40bd376cd6cd3
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>