openembedded-lkft: generic/developer: don't use lkftmetadata class
The lkftmetadata.bbclass has lots of advantages, a just a few
drawbacks that are, unfortunately, very annoying

One is that it requires the clean-up of the important packages
on every build, so that the metadata can be generated again.
This means that those packages have to be rebuilt even if they
or their dependencies have not changed at all.

Another problem is that Bitbake then may fail to find the info
it needs for the creation of packages. Every now and then the
builds might fail with:
  Function failed: sstate_create_package

Change-Id: Ia6bde9fb0d8186b05eaf3b6394b75a7d47fdfe2a
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>