openembedded-lkft-linux-developer: set missing variables for testing
A few environment variables are still needed for the test
templates. These are generally set in regular jobs because
there is a kernel checkout in the trigger itself, which is
able to determine and pass the information about the git
version (KERNEL_DESCRIBE) and the version reported by the
codebase via the main Makefile (MAKE_KERNELVERSION).

For now, we just go without them. We need here a checkout
of the kernel so that we can ourselves determine the value
of those variables.

Change-Id: Ibd5e8da1c0f3b25cf2f45450e4f1532c3ded0854
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>