openembedded-lkft-linux-developer: process specific test suites to run
A specific test suites can now be given as a parameter
(TEST_SUITES) so that if the test suite (.yaml) exists, then
it is correspondingly added to the list of test suites to run.

The behavior changed for the case of "none": now it simply
terminates execution of the current shell, so that nothing
continues. From there on, anyway, everything's related to the
tests. This avoids a problem with running submit_for_testing
without any test plan, which would make the build break (and
didn't make much sense).

Change-Id: I7c841e6791880e58d7d54cb13a58eaaa74defabd
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>