openembedded-lkft: groovy: read optional job name
When supported by submit_for_testing.py, the job name will be
printed along with the URL of the test job. The pattern
matcher needeed a slight change to correctly detect that line.

If a job name is printed, it will be added to the link to the
test job. Exempli gratia:

Whereas before links (within <a/> tags) looked like this:
  QA Reports: 266596
now they will look like this:
  QA Reports: 266596 lkft-ltp-syscalls-master-40

If no job name is printed, nothing changes in the link
compared to the existing situation.

Change-Id: Iee4b0393f1fba5d5043f299503df85e43524203d
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>