lt-qcom-openembedded: Remove long run LAVA test jobs for now
We are experimenting high test loads in our LAVA LAB's one reason is the
long time consuming testing (stress, ptest, piglit, gst-validate) the
other one is the lack of enough boards to cover the test load.

We want to have enough resources to be able do a fast boot regression
testing in our kernel integration branch.

For now remove the time consuming test jobs, next are the execution
times in dragonboard-410c:

ptest takes ~90 minutes
stress takes ~270 minutes
piglit takes ~180 minutes
gst-validate ~180 minutes

At some point in the future we want to execute that testing but not in
an every build basis.

Change-Id: I6bdf7f8aba4edfbd7af459e86b3e1755a99eea57
Signed-off-by: Aníbal Limón <anibal.limon@linaro.org>