LKFT Developer build email notifications
Initial jobs to provide developer build email notifications.

trigger-lkft-notify-developer runs every 5 minutes and looks in
qa-reports for new builds that have finished testing, but that have not
yet been notified. If it finds any, it calls lkft-notify-developer with
the build_url. lkft-notify-developer, in turn, sends a notification to
the developer based on a metadata field in the build.

Change-Id: Ib29fabc124fbe60362c985ad4132f43c44c0ed90
Signed-off-by: Dan Rue <dan.rue@linaro.org>