Fix here-doc indentation in tcwg-regression-detection
Delimiter of the here-doc for LLVM release regression detection is
idented as per the start of the here-doc. However, here-docs require
delimiter to not have any indentation in a script. This patch changes
the indentation in the jenkins job for the delimiter to have no
indentation relative to the bash script.

Change-Id: Ibc6972020a020ad60524f596d3f6e99d16b8fe3e