lt-qcom-linux-aosp: Disable bootimage packaging and boottesting for DB410c
DB410c bootimage packaging and testing are not functional at
the present moment. We will switch to DB820c anyway when it
is ready for testing so for now disable the DB410c bootimage
packaging and testing after kernel build.

The bootimage packaging and submit-for-testing scripts,
templates and variables are not removed now. When DB820c is
ready for testing we will either use them again or move the
bootimage creation to the AOSP build job and then remove them
from this kernel job.

Change-Id: I9217d420d261cee18350b1da283c4d9a53273f36
Signed-off-by: Todor Tomov <todor.tomov@linaro.org>