lt-qcom-linux-aosp: Fetch android config fragments
The android kernel config fragment files are stored on a
separate googlesource repo. A script to fetch them is present
in the kernel source. Before doing the kernel config, run the
script to fetch the config fragment files.
Publish the config fragment files along with the resulting files
from the build to keep trace of the config used.

Change-Id: I6fac19ba6222a504478874469e1d96a3a0023eac
Signed-off-by: Todor Tomov <todor.tomov@linaro.org>