tcwg-*: Switch to email-ext, set 'aborted' to true.
'email' does not send an email when a build was aborted, which made us
miss some infrastruture problems.

This patch replaces 'email' with 'email-ext' in all tcwg* jobs using
'email', and adds 'aborted: true'. In some cases where we had both
'email' and 'email-ext', use only the latter as it is more
flexible/powerful.

Do not add 'aborted: true' in the jobs that already have 'always: true'.

Change-Id: I60a8dc287216bd1eb5705690505928f85fcf4815