lt-qcom-linux-testimages: add new job
This is a new job to create testimages for QC LT kernel testing. So
far, we were hijacking the lt-qcom-openembedded-rpb-sumo job which is
meant to be used for Dragonobard OE releases. We now create a proper
job for Kernel CI so that we can add more images/features without any
impact on OE releases.

That should also make everything simpler when we need to create
testimage for private boards, or boards that are not supported by
meta-qcom.

The main changes with the existing test images we used so far are:
* build generic images for qemuarm (32bit) and qemuarm64 instead of an
image for a specific board. Our kernel testing user space is mostly
board agnostic.
* export all files to linaro.org, in member-builds/qcomlt instead of
96boards..

Change-Id: Id7705f6025f97ab862aab1fd68a1b8d2515cbd35
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>