lt-qcom-linux-integration,mainline: use DTB name for MACHINE
Insted of using commercial board name (e.g. Dragonboard 410c, ...)
let's use the DTB name, since we are mostly dealing with kernel
testing here, and the DTB name should be the identifier.

Sometimes we use MACHINE in bash/shell variables, so we need to make
sure that we don't use '-' character, and in these situations it is
replaced with '_' instead.

Change-Id: Icec0e60a0ee9fae47f962b6a652bbe0965995e32
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>