tcwg_kernel*: Reduce disk usage of "build" and "bisect" jobs
At the moment we use per-configuration workspace for each of
128 configurations of tcwg_kernel build and bisect jobs.  This
patch makes configurations of same toolchains share the workspace.

To avoid parallel builds using the same workspace we include
$EXECUTOR_NUMBER is the workspace path.

Re. "trigger" jobs, it's not entirely clear whether we can share
workspaces among them.  Jenkins may or may not rely on contents of
cloned repos not changing between SCM polls.  Since these jobs
run on tcwg-coordinator label, and their workspaces are small, we
keep them as-is for now.

Change-Id: Ia0d3e24da499e3b46f58e9eaa86b8019cd5f8fbe