openembedded-lkft: Use GCC 7.3
The compiler that we have in RPB, linaro-gcc 7.1, is not going
to cut it anymore due to recent configurations in the kernel
that require the retpoline mitigation to be in place.

Example problem during configure:
  NOTE: make -j 8 -C /srv/oe/build/tmp-rpb-glibc/work/intel_core2_32-linaro-linux/linux-generic-next/4.19+gitAUTOINC+dfe048f91e-r0/git O=/srv/oe/build/tmp-rpb-glibc/work/intel_core2_32-linaro-linux/linux-generic-next/4.19+gitAUTOINC+dfe048f91e-r0/build kselftest-merge
  make: Entering directory '/srv/oe/build/tmp-rpb-glibc/work-shared/intel-core2-32/kernel-source'
  make[1]: Entering directory '/srv/oe/build/tmp-rpb-glibc/work/intel_core2_32-linaro-linux/linux-generic-next/4.19+gitAUTOINC+dfe048f91e-r0/build'
  arch/x86/Makefile:224: *** You are building kernel with non-retpoline compiler, please update your compiler..  Stop.
  make[1]: Leaving directory '/srv/oe/build/tmp-rpb-glibc/work/intel_core2_32-linaro-linux/linux-generic-next/4.19+gitAUTOINC+dfe048f91e-r0/build'
  Makefile:152: recipe for target 'sub-make' failed
  make: *** [sub-make] Error 2
  make: Leaving directory '/srv/oe/build/tmp-rpb-glibc/work-shared/intel-core2-32/kernel-source'
  ERROR: oe_runmake failed

This has been failing in linux-next since next-20181121, and
has recently started to fail on torvalds/master (mainline).
Now that such requirement has been merged into mainline, it's
also been backported to other trees, like stable-rc 4.14 and
4.19. We expect all trees to fail from now on.

So, instead of using:
  GCCVERSION = "linaro-7.%"
we now set it to:
  GCCVERSION = "7.%"
at the distro level, which will give us GCC 7.3.0 as packaged
by OpenEmbedded (core).

Change-Id: Ia6055a145ca4a30aa4ec9a8e57921d153db5a192
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>