openembedded-lkft: Fix pub. dest. due to manifest name
The publishing destination used to contain the (correct)
manifest branch, which is "rocko", but that was recently
changed to "lkft/rocko", which buried the artifacts one level
deeper in the snapshots directory structure.

This change removes the "lkft/" part from the $MANIFEST_BRANCH
variable of the publishing destination, so that the artifacts
are published where they used to.

Change-Id: I6668a30df7a98f857cb65e4f0b2e81b9a3401eea
Reported-by: Anders Roxell <anders.roxell@linaro.org>
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>