openembedded-lkft: developer: Use LKFT distro
The LKFT distro is now used throughout the stable RC,
-mainline and -next branches, and is working better than
ever.

The Developer Build still needed to be migrated to it, so
now's a good time for this. This requires a few small changes
in the builders.sh script so that it accomodates the
SRCREV_kernel variable, used elsewhere (instead of
KERNEL_COMMIT). The bits/wordsize.h part does not apply
anymore. Most of the cruft due to using RPB as a distro is
now if'ed out, as in the regular jobs.

Change-Id: I3c409645c7e80ab7670714af89e2cc2c3165eb36
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>