lkft, openembedded-lkft: generate SHA256SUMS.txt for deployed files
As part of the LSS-423 effort, generate SHA256SUMS.txt on the CI side so future LAVA jobs can they have the correct and uncorrupted files for testing.

Debian marks 'coreutils' as 'essential', which means it should already be present, but other CI jobs explicitly request it, so take the belt-and-suspenders approach and add it to the list of packages to be installed.

Change-Id: I1087ab05e809061d48f7690b16a845f5096215d0
Signed-off-by: Koen Kooi <koen.kooi@linaro.org>