tcwg-benchmark: Stop sending emails
Now that tcwg-buildslave is triggerring tcwg-benchmark jobs
(for tcwg_bmk project) we have emails going to tcwg-buildslave@linaro.org
LDAP group for every build.

It was unfortunate for us to name LDAP groups tcwg-buildslave and
tcwg-benchmark the same as automation users tcwg-buildslave and
tcwg-benchmark, but there is no easy way around that.  So stop
sending email to requester.

Change-Id: Id673235a5b2fd7e7b601fec5a213d795931814cf