tcwg-cleanup-stale-workspaces: Fix job waiting in the queue
Due to TCWG nodes being busy all the time the cleanup job waits
indefinetely in the queue.  Our goal is to avoid starting new jobs
on a node that has cleanup running.  Occupying all executors should
achieve this.  We have at most 10 executors (on coordinator nodes),
so set weight to 10.

Change-Id: Ie4008912c0138cf87b6eb13d2fb438f41592e80b