tcwg-llvm-release: Reduce parallelism to 3 on TK1s
Using all 4 cores for linking eats up too much RAM, so we get a lot of
swapping (and potentially timeouts) when running the release job. Since
the release script does not support Ninja yet, we cannot limit only the
number of link jobs. Therefore, we hardcode a lower number of build jobs
as a stop-gap solution. On the buildbots, we use 4 jobs in general and 2
for linking. Hopefully, a value of 3 is a reasonable compromise between
those values.

Change-Id: Iaf067a904f51699d8ba57d34d0221be41b1e47ad