trigger-lt-qcom-linux-integration: fix location of kernelci.org files
Files published by kernelci.org are now in the 'gcc-7' subfolder. This
patch is just a quick hack to get things going. But we should
definitely revisit how we parse the kernelci folders to find artifacts
to tests, as it's likely to become gcc-8 someday and break again!

There is build.json file that could be used , but it's located in the
gcc-7 folder, so we need something smarter sooner or later...

Change-Id: I1e044aa598a9fe46ba64742f457c954807e3f91d
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>