lt-qcom-linux-{integration,mainline}: simplify list of parameters
We recently started to use generic images for all boards, to ROOTFS
and RAMDISK variables don't need to be per-board, they are all the
same.

Also we can assume that all DTBS are located under the same folder, so
we can make KERNEL_DT_URL generic and indicate a folder where all DTBS
can be found. We are still checking that DTBS for all supported
MACHINES exist.

This patch also removed all unused (or uninitialized) job parameters,
and we set MACHINES in the top level .yaml file to better control the
list of MACHINES to test for each job.

Change-Id: If0b987e83b429f8bca8f1192c1f8330268782a5e
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>