tcwg-regression-detection: Modernize LLVM release job
Launch the LLVM release job on a xenial container and for LLVM 7.0.0
instead of 6.0.0.

Note that we are diverging from using $distro, like the rest of the job,
but we have agreed in the LLVM sync that this is acceptable and we
should in fact split the whole thing into its own independent job
someday (jenkins-scripts-regression-detection?).

Change-Id: I564e963625cbfbcbf161eb2cb16e28eec43eaea2