tcwg-llvm-release: Use D05 for AArch64 release
Make it possible to use the D05 for the AArch64 release and set it as
the default platform. This is precipitated by PR40511, which is
reproducible on the APM containers, but not on the D05 containers, since
they use a different file system.

I'm not sure how many buildjobs we want to use, I have run this
successfully with 64. Since the containers are prioritized, this
shouldn't hog the whole machine and make it unusable.

Change-Id: Icb0b6aa0b45d1e8d85f11e1331765b87ea4cf004