tcwg-llvm-release: Change default scripts branch
At the moment we are using 'llvmprod' as the default branch for the
Jenkins scripts repo. However we never had a good process in place for
updating that branch, so it often lagged behind and we ran into all
sorts of fixed-on-master issues during the release.

This patch changes the default branch to 'tested', which is updated
automatically by the TCWG regression detection job.

Change-Id: Ic1b9a1367404b910820d31875fea6100bebb1eca