Parse PARTLABEL bootarg to determine rootfs partition
In order to make the handling of rootfs partitions more generic and more
flexible, we introduce a bootarg PARTLABEL that tells us what partition label
should be used as the rootfs partition.

If no PARTLABEL bootarg is found, we use the default for the specific board we
are building things for.

Our custom init script will take care of mounting the rootfs partition.

Change-Id: If52cbec0800bde86057668b34f7ea0f7fc67154a
Signed-off-by: Luis Machado <luis.machado@linaro.org>