openembedded-lkft: Workaround for Juno/4.4 and juno-r2.dtb
There is no juno-r2.dtb, and when using the generic Linux
recipe it's not possible to account for the missing DTB via an
override. For example, the stable-rc_4.4.bb recipe says:
  KERNEL_DEVICETREE_remove_juno = "arm/juno-r2.dtb"
but removing it from the machine definition in the generic
recipe is not as straight forward.

This works around that limitation, thus enabling Juno and 4.4
for the generic builds.

Change-Id: Ibfc6a9feee7aab3e6e108430c86377021f7eccfd
Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>