Merge luajit-aarch64-perf into luajit-aarch64
The luajit-aarch64 project has not been triggered for about 2 years
and it makes more sense for the luajit-aarch64-perf job to be merged
into it instead.

Going forward (i.e. in the next few days), I intend to make this into
a matrix project so that we can add more architectures to it.  While
the benchmark is being exercised, they're pretty useless given that
they're running on VMs and the scaffolding for the benchmark is
rudimentary.

Change-Id: Ie5b75759249d2bfc3219c0de5b2327cdd2f287ce