luajit: Add job for my github fork
I have made a bunch of changes to the luajit project, mainly:

- Fixed a register allocator bug that affected arm64
- Improved XZR use
- Added scaffolding to run and measure benchmarks
- Added `make check` and `make bench` targets for tests and benchmarks
- Merged the LuaJIT-test-cleanup repo into mainline

The changes have not been accepted in the official repo but then the
official repo seems pretty inactive with Mike Pall only occasionally
fixing some bugs on his own volition and rarely providing feedback on
PRs or issues.  I have also sent emails to the mailing list and was
met with silence.  As a result, it made sense to me to keep the fork
in a maintainable state so that if upstream does wake up some day (or
if there is a change of heart in the community) then this body of work
will remain useful.

Change-Id: Icafe80d354df275f28a71d03d137648f11df18d0