Fix warning regarding timeouts in in job definition
See https://lkft.validation.linaro.org/scheduler/job/674690#L0

Not any action block timeout could be higher than the global job timeout.
This patch is not changing the current behavior as the timeout is roughly the
minimum of the job timeout and the action timeout.

In the current LAVA version, this is only a warning.

Change-Id: I49379fa7f56a8be7473a5531b82e0ae3f55e098f